+++ Get-LaptopSoftpaq.ps1
@@ -15,12 +15,15 @@ function Get-LaptopSoftpaq {
     $progressPreference = "SilentlyContinue"
 
     $Bios = (Get-ComputerInfo).BiosManufacturer
-    $CmslPath = Test-Path -Path "C:\Program Files\WindowsPowerShell\HP.CMSL.UninstallerData"
     if ($Bios -eq "HP" -and $CmslPath -notmatch "True") {
-        Invoke-WebRequest -Uri "https://hpia.hpcloud.hp.com/downloads/cmsl/hp-cmsl-1.6.7.exe" -OutFile "C:\Windows\Temp\HpModule.exe"
-        Start-Process -FilePath "C:\Windows\Temp\HpModule.exe" -Wait -ArgumentList "/verysilent /norestart"
-        Remove-Item -Path "C:\Windows\Temp\HpModule.exe" -Force
-        #Start-Process -FilePath "https://developers.hp.com/hp-client-management/doc/client-management-script-library"
+
+        $CmslPath = Test-Path -Path "C:\Program Files\WindowsPowerShell\HP.CMSL.UninstallerData"
+        if ($CmslPath -notmatch "True") {
+            Invoke-WebRequest -Uri "https://hpia.hpcloud.hp.com/downloads/cmsl/hp-cmsl-1.6.7.exe" -OutFile "C:\Windows\Temp\HpModule.exe"
+            Start-Process -FilePath "C:\Windows\Temp\HpModule.exe" -Wait -ArgumentList "/verysilent /norestart"
+            Remove-Item -Path "C:\Windows\Temp\HpModule.exe" -Force
+            #Start-Process -FilePath "https://developers.hp.com/hp-client-management/doc/client-management-script-library"
+        }
     
         $TestPath = Test-Path -Path "C:\SOFTPAQ"
         if ($TestPath -match "False") {
